+++ apps/showcase/lib/demos/input_additional_demo.dart
@@ -28,7 +28,7 @@ class _InputAdditionalDemoState extends State<InputAdditionalDemo> {
   Widget build(BuildContext context) {
     final ui = UiTheme.of(context);
 
-    return SizedBox(
+    return UiBox(
       width: 400,
       child: VStack(
         crossAxisAlignment: CrossAxisAlignment.start,