+++ apps/showcase/lib/demos/ui_box_demo.dart
@@ -9,7 +9,7 @@ class UiBoxDemo extends StatelessWidget {
   Widget build(BuildContext context) {
     final ui = UiTheme.of(context);
 
-    return SizedBox(
+    return UiBox(
       width: 400,
       child: VStack(
         crossAxisAlignment: CrossAxisAlignment.start,